requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
schedule>=1.2.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
python-telegram-bot>=20.0
matplotlib>=3.7.0
